+++ .github/workflows/rust.yml
@@ -33,7 +33,7 @@ jobs:
                 "this.job == 'lint'": [stable, nightly]
                 "this.job == 'test-used-linker'": nightly
                 "this.job == 'build'": [stable, beta, nightly]
-                "this.job == 'build-minimal'": ["1.85"]
+                "this.job == 'build-minimal'": ["1.85", "1.88"]
                 "this.job == 'build-ctor-minimal'": ["1.65"]
                 "this.job == 'wasm'": [stable, "1.85"]
                 "true": "stable"
