+++ link-section/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.19.3] - 2026-08-09
+
+### Fixed
+
+- Fix section pinning on OpenBSD. (https://github.com/mmastrac/linktime/pull/509)
+
 ## [0.19.2] - 2026-07-29
 
 ### Changed
