+++ tests/link_section/copied/copied-types/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "copied-types"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+link-section = { path = "../../../../link-section" }
