+++ tests/link_section/copied/copied/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "copied"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+copied-types = { path = "../copied-types" }
+register-a = { path = "../register-a" }
+register-b = { path = "../register-b" }
+
+[[bin]]
+name = "copied"
+path = "src/main.rs"
