+++ tests/link_section/copied/register-b/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "register-b"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+copied-types = { path = "../copied-types" }
+link-section = { path = "../../../../link-section" }
